:recycle: (refactor) move send message interpretation to interpreter
