+++ libs/functions/messages/whatsapp/src/lib/models/send-message-main.model.ts
@@ -1,5 +1,6 @@
+import { BaseChannel } from "@app/model/bot/channel";
 import { BaseMessage } from "@app/model/convs-mgr/conversations/messages";
-import { StoryBlock, StoryBlockTypes } from "@app/model/convs-mgr/stories/blocks/main";
+import { WhatsAppBaseMessage } from "@app/model/convs-mgr/functions";
 
 /** 
  *  @Descripion: Abstract Class for sending messages 
@@ -8,8 +9,5 @@ import { StoryBlock, StoryBlockTypes } from "@app/model/convs-mgr/stories/blocks
 export abstract class SendMessageModel {
     
     /** Checks the type of message and calls the appropriate method */
-    abstract sendMessage(message: BaseMessage, endUserPhoneNumber: string, block?: StoryBlock,): Promise<any>
-
-    /** Converts the message to text based on platform api docs and sends it */
-    protected abstract _sendTextMessage(message: BaseMessage, endUserPhoneNumber: string, block?: StoryBlock,): Promise<any>
+    abstract sendMessage(message: BaseMessage, channel: BaseChannel): void
 }
